Move etreeIterWithDepth to a higher level for reuse by other jurisdictions
