+++ arelle/plugin/validate/ESEF/Util.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 import json
 import os
 from collections import defaultdict
-from collections.abc import Collection, Generator
+from collections.abc import Collection
 from typing import Any, Union, cast
 import regex as re
 
@@ -177,13 +177,3 @@ def _hasEventAttributes(elt: Any, attributes: Collection[str]) -> bool:
     if isinstance(elt, _Element):
         return any(a in attributes for a in elt.keys())
     return False
-
-
-def etreeIterWithDepth(
-    node: ModelObject | _Element,
-    depth: int = 0,
-) -> Generator[tuple[ModelObject | _Element, int], None, None]:
-    yield node, depth
-    for child in node.iterchildren():
-        for n_d in etreeIterWithDepth(child, depth + 1):
-            yield n_d
